scripts/SubsetSelection/requirements.txt
h5py
torch
numpy
datasets
accelerate
transformers
evaluate
peft
accelerate
submodlib-py==0.0.3
tqdm
jinja2
#faiss-gpu
faiss-cpu
langchain[openai]
python-dotenv
tenacity
sentence-transformers
kneed
einops
